# docs/swagger.yaml
basePath: /api/v1/
definitions:
  model.Note:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
  request.CreateUpdateNoteRequestDto:
    properties:
      content:
        type: string
    required:
    - content
    type: object
host: localhost:8000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Fiber Rest API
  version: "1.0"
paths:
  /notes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Note'
            type: array
        "400": {}
      summary: Returns a 200
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUpdateNoteRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Note'
        "400": {}
      summary: Returns a 200
    put:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Note'
            type: array
        "400": {}
      summary: Returns a 200
  /notes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204": {}
        "400": {}
      summary: Returns a 204
    get:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Note'
        "400": {}
      summary: Returns a 200
swagger: "2.0"